refactor(frontend): Extract component AddTokenReviewByNetwork
